requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sentence-transformers
numpy
faiss-cpu
pytest
httpx
pytest-asyncio
pydantic-settings
google-genai
